Reeformat health checks and restart if fails
