+++ src/main/kotlin/com/fcfb/discord/refbot/model/health/Health.kt
@@ -1,0 +1,10 @@
+package com.fcfb.discord.refbot.model.health
+
+data class Health(
+    val status: String,
+    val jobs: Map<String, Boolean>?,
+    val memory: Map<String, String>?,
+    val diskSpace: Map<String, String>?,
+    val kord: Map<String, String>?,
+    val message: String? = null,
+)
